Council audit: security hardening, Decimal precision, cleanup
Backend:
- float → Decimal on all financial model fields (payroll settings, tax brackets, deductions, employee salary)
- Mass assignment whitelist on payroll service setattr() calls
- employee_ref: global unique → compound (company_id, employee_ref) via migration 005
- Add missing company_id indexes on locations, accounts, employees, reference_data (migration 005)
- CORS: restrict allow_methods and allow_headers to explicit list
- Language endpoint: dict → LanguageUpdate Pydantic schema
- Fix alembic.ini URL for Docker (127.0.0.1 → db)

Frontend:
- Add React ErrorBoundary component wrapping entire app
- Add ApiError type + getErrorMessage helper in api.ts
- Fix TypeScript error in PayrollSetup (LeaveType cast)

Cleanup:
- Move demo HTML files to docs/designs/
- Move portal screenshots to docs/screenshots/
- Update CLAUDE.md with audit results, known issues, backend gotchas

+++ backend/app/models/payroll/deduction_type.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+from decimal import Decimal
 from typing import Optional
 import uuid
 
@@ -24,7 +25,7 @@ class DeductionType(Base):
     calculation_method: Mapped[str] = mapped_column(
         String(30), default="FIXED"
     )  # FIXED, PERCENTAGE_GROSS, PERCENTAGE_BASIC, TAX_TABLE, MANUAL
-    default_value: Mapped[Optional[float]] = mapped_column(Numeric(15, 2))
+    default_value: Mapped[Optional[Decimal]] = mapped_column(Numeric(15, 2))
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     description: Mapped[Optional[str]] = mapped_column(Text)
     sort_order: Mapped[int] = mapped_column(Integer, default=0)
